extract BaseEditorPanel.SelectionHighlightHandler -> EditorHighlightHandler
